+++ src/main/java/com/irrigation/erp/backend/dto/StockAdjustmentRequestDTO.java
@@ -4,13 +4,15 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 public class StockAdjustmentRequestDTO {
 
     // Getters and Setters
     @NotNull(message = "Quantity change cannot be null")
-    private Double quantityChange;
+    private BigDecimal quantityChange;
 
     @NotNull(message = "Adjusting user ID cannot be null")
     private Long adjustingUserId;
